requirements.txt
asn1crypto==1.5.1
iniconfig==2.0.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pg8000==1.31.2
pluggy==1.5.0
pyarrow==18.0.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scramp==1.4.5
six==1.16.0
tzdata==2024.2
